feat: novo controller para progresso de habito
