+++ backend/controllers/ProgressController.js
@@ -1,0 +1,69 @@
+const Progress = require("../models/Progress")
+const Habit = require("../models/Habit")
+
+class ProgressController {
+    async habitProgressCreate(request, response) {
+        var {habit_id, name, type, parameter} = request.body
+        habit_id = Number(habit_id)
+        type = Number(type)
+        parameter = Number(parameter)
+        if (!name || name.trim() === '') { 
+            response.status(400)
+            response.json({err: "nome invalido."})
+            return
+        }
+        if (isNaN(habit_id)) {
+            return response.status(400).json({ err: "Hábito invalido." })
+        }
+
+        if (isNaN(type)) {
+            return response.status(400).json({ err: "Tipo invalido." })
+        }
+
+        if (isNaN(parameter)) {
+            return response.status(400).json({ err: "Parâmetro invalido." })
+        }
+
+        var valid = await Habit.habitExist(habit_id)
+        if(!valid) {
+            response.status(404)
+            response.json({err: "Nenhum habito encontrado."})
+            return 
+        }
+        var nameExists = await Progress.findNameProgress(name)
+        if(nameExists) {
+            return response.status(406).json({err: "nome de progresso ja existe."})
+        }
+        var done = await Progress.newHabitProgress(habit_id, name, type, parameter)
+        if(done) {
+            response.status(200)
+            response.send('Progresso de habito adicionado com sucesso.')
+        }
+        else {
+            response.status(500)
+            response.json({err: "erro ao cadastrar progresso de habito."})
+        }
+    }
+
+    async getHabitsProgress(request, response) {
+        const habit_id = request.params.habit_id
+        if (!habit_id || isNaN(habit_id)) {
+            return response.status(400).json({ err: "Usuário inválido." });
+        }
+
+        try {
+            const habit_progress = await Progress.getAllProgressHabits(habit_id)
+            if(habit_progress != undefined) {
+                response.status(200).json({habit_progress: habit_progress})
+            }
+            else {
+                response.status(404).json({ err: "progressos não encontrados."})
+            }
+        } catch(err) {
+            response.status(500).json({ err: "Erro ao buscar progresso de habitos."})
+        }
+    }
+
+}
+
+module.exports = new ProgressController()
